Wrap `RobotDriver` to fix pure virtual method call in destructor.
With the original design, the `shutdown()` method is pure virtual at the
moment it is called in `~RobotDriver` (at this stage the implementation
of the derived class is not visible as the derived class is already
destroyed).

Resolve this by wrapping `RobotDriver` in a class `MonitoredRobotDriver`
which handles the timing monitoring (i.e. if actions are applied fast
enough).  In the actual `RobotDriver` class only the basic pure virtual
functions remain.  The call of the `shutdown()` method is done from the
destructor of `MonitoredRobotDriver` and therefore happening before
`RobotDriver` is destroyed.