Remove Status from NJointRobotTypes
Apparently pybind11 has problems with creating multiple bindings of the
same type.  Since `Status` is not templated anymore, it was exactly the
same in all "types" modules.  This compiled without error but when
importing two of the resulting modules (as happening in the
`__init__.py`), it raised the following error: "ImportError:
generic_type: type "Status" is already registered!".

To fix this, move the binding of `Status` to a new module `py_generic`.

To be consistent between Python and C++ remove the typedef from
`NJointRobotTypes`.

Add bindings for the constructor of `Status` and `Observation` so they
can be instantiated in Python.