Add function process_desired_action
Add a function `process_desired_action()` which contains the processing
of the action (position control and safety-checks) which was originally
implemented in the `NJointBlmcRobotDriver` class in blmc_robots.

The purpose of moving it here in a separate function is to be able to
use it in other places as well and to be more testable.

For a cleaner interface wrap the function in a templated struct
`NJointRobotFunctions` similar to the `NJointRobotTypes`.