Split RobotData into single and multiprocess versions
The `RobotData` class remains as base class, holding the pointers to the
time series, but does not initialize them.  Initialization is done in
the derived classes `SingleProcessRobotData` (using the single process
time series) and `MultiProcessRobotData` (using the multi process time
series).

This split is needed because the originally intended design with a
if-check does not compile with template types that are not serializable,
even when only single process is used.

While doing this, also remove the `Status` argument from the templates
(the type for this is fixed anyway).