Enable usage of multi-process time series in RobotData (#36)
* Split RobotData into single and multiprocess versions

The `RobotData` class remains as base class, holding the pointers to the
time series, but does not initialize them.  Initialization is done in
the derived classes `SingleProcessRobotData` (using the single process
time series) and `MultiProcessRobotData` (using the multi process time
series).

This split is needed because the originally intended design with a
if-check does not compile with template types that are not serializable,
even when only single process is used.

While doing this, also remove the `Status` argument from the templates
(the type for this is fixed anyway).

* Cleanup CMakeLists.txt

Remove commented and unnecessary stuff

* Increase timing thresholds in demo.cpp

As this demo does not use a real robot, it can safely be run on
non-real-time systems.  However, on those systems the original
thresholds can be too tight, so increase them by a significant factor to
ensure the demo runs without unexpected errors.

* Add minimalistic multi-processing demo

This is based on demo.cpp but split into separate executables for
backend and frontend, using the `MultiProcessRobotData` for
communication.

* Move cereal code for Eigen to separate file

This should eventually be moved to some "serialization tools" package
but this does not yet exist, so keep it here for now.  However, move it
to a separate file due to different license.