Accessing a single camera using the Pylon interface (#41)
Adds a pylon driver based on the pylon Camera Software Suite from Basler to interact with cameras that depend on it. Enables grabbing the latest images from the camera stream and appending them to a time-series of camera observations, consisting of the frame and the corresponding timestamp of capture. The interface is exactly the same for both the opencv and the pylon drivers. The user can switch between the two by just specifying the driver to be created.