Add wait_until_terminated to RobotBackend
Add a method `wait_until_terminated` to the `RobotBackend` that blocks
until the background loop terminated.  This is useful for
backend processes in a multi-process setup where the backend is started
and then the process should loop until the backend is terminating.

Use the new GlobalSignalHandler to add SIGINT as a termination
criterion.