Add "real-time mode" option to RobotBackend
Use an explicit "real_time_mode" flag to decide if the backend should
wait for the next action or repeat the last one if it is not provided in
time.

This replaces the check for finity of the action duration timeout as
there are use cases in simulation where action repetition is desired but
enabling the timeout is not.