Do not create monitored driver inside backend
Instead of always wrapping the robot driver in a monitored driver, just
use the driver as it is given.  The user is now responsible to pass a
monitored driver in case the monitoring is desired.  This has the
advantage that for cases where no monitoring is desired (e.g.
simulation), no wrapper with infinite timeouts needs to be created, but
the actual driver can just be used directly.