Simplify template of MonitoredRobotDriver
Instead of again specifying Action and Observation in the template of
the MonitoredRobotDriver, only put the type of the Driver that is
wrapped.

In the base RobotDriver class add typedefs for the templated Action and
Observation, so they can be used here to specify the types.

With this design, starting from C++17, it is even not needed anymore to
specify any template argument for MonitoredRobotDriver as it can be
deducted from the passed driver pointer.