Syncing Three pylon Dependent Cameras (#47)
Adds a TriCameraDriver that creates three instances of the PylonDriver and returns the combined observation from them in the form of an array of three CameraObservations. Grabbing from the three cameras can be performed almost synchronously: the max difference between the timestamps (corresponding to ~ the instant when a frame is grabbed) of the observations returned from any two cameras is of the order 1e-3.