Release the GIL in the constructor
It is important to release the GIL when destructing the backend, as
otherwise the program will get stuck in a dead lock in case the driver
needs to run some Python code (as is the case with the pyBullet driver).

Unfortunately, there does not seem to be a non-intrusive way to do this
in the destructor when using pybind11.