Implement one-shot logging function
Implement method `write_current_buffer()` which writes content from the
current time series buffer to a log file without running a thread in the
background.  Use this to log data if the time span is short enough to be
fully contained in the time series buffer.

Also add a flag to indicate if the thread is running so that, for
example, the `stop()` can detect if it is called without the thread
running.