Make Status::error_message fixed-size
For use in a multi-process time series, the `Status` objects need to
have fixed size.  Thus replace the variable-size `std::string` with a
fixed size `char` array.  The length is set to 64 characters which is
more than enough for our current error messages (typically < 40 chars).

Since working with plain char-arrays is pretty annoying, make it private
and only access it through methods that convert the message from/to
`std::string`.