Add buffer-based logging like in SensorLogger
Add a logging method using `start()` and `stop_and_save()` which uses an
background loop to copy data to an internal buffer (same method as in
`SensorLogger`).

This has the advantage that it does not depend on all relevant data
being held in the robot data time series.

This method supports binary or CSV files, the latter plain or with gzip
compression.