fix: Do not block sensor logger if there is no data
Instead of blocking and waiting for the oldest time index, simply start
with `t = 0` if there is no data yet in the beginning.  The same fix has
already been applied recently on the RobotLogger class and prevents the
`stop()` method from blocking forever in case the logger is stopped
before any data is received.