Indicate shortened error messages
If an error message passed to Status::set_error exceeds the maximum
length Status::error_message can hold (currently 64 chars including
terminating \0), it simply gets cut.  This can be a bit confusing.  To
hopefully make it a bit more understandable, replace the last character
in the string with `~` in this case (unfortunately more appropriate
characters like … are not available with standard ASCII).