Add optional sensor_info to sensor interface
Add a `sensor_info` field to `SensorData` which can be used to provide
static information about the sensor to the user.  A method
`get_sensor_info()` is added to `SensorDriver` to fill the field in the
backend.  It can then be accessed via a method of the same name in
`SensorFrontend`.

By default, it provides an empty struct for backward compatibility.

Use this for data that doesn't change between observations (e.g.
calibration parameters of frame rate).  Changing values should be added
to the observation instead.